Refactor functions a bit
